Revert "refactor: modify review item component"
