+++ main.js
@@ -1,10 +1,6 @@
 import "./components/button-link.js";
 import "./components/footer-link/footer-link.js";
 import "./components/header.js";
-import "./components/hero/hero.js";
-import "./components/link-button/link-button.js";
-import "./components/review-item.js";
-import "./components/button-link.js";
 import "./components/hero.js";
 import "./components/image.js";
 import "./components/review-item/review-item.js";
